fixing when participant null abort send message bot

+++ send/telegram.ts
@@ -12,24 +12,21 @@ const CHAT_ID = Number(config.get('chat.id'))
 const TELEGRAM_BOT = config.get('telegram.bot')
 
 const sendTelegram = async (payload: bodyInterface): Promise<void> => {
-  try {
-    const picture = await captureScreenshot(payload.url)
-    const messageByCreated = formatByCreated(payload)
-    const messageByReview = formatByReview(payload)
-    if (!picture) {
-      sendMessageWithBot(TELEGRAM_BOT, CHAT_ID, messageByCreated)
-      sendMessageWithBot(TELEGRAM_BOT, CHAT_ID, messageByReview)
-    } else {
-      const messageId = await sendPhotoWithBot(TELEGRAM_BOT, CHAT_ID, picture)
-      telegramSendUser.add({ chat_id: CHAT_ID, message: messageByCreated, message_id: messageId }, gitOptions)
-      telegramSendUser.add({ chat_id: CHAT_ID, message: messageByReview, message_id: messageId }, gitOptions)
-      fs.unlinkSync(picture)
-    }
-    sendBodyIsValid(payload)
-  } catch (error) {
-    console.log(error.message)
-    throw error
+  const messageByCreated = formatByCreated(payload)
+  const messageByReview = formatByReview(payload)
+  const picture = await captureScreenshot(payload.url)
+  const messageId = await sendPhotoWithBot(TELEGRAM_BOT, CHAT_ID, picture)
+
+  if (picture && messageId) {
+    telegramSendUser.add({ chat_id: CHAT_ID, message: messageByCreated, message_id: messageId }, gitOptions)
+    telegramSendUser.add({ chat_id: CHAT_ID, message: messageByReview, message_id: messageId }, gitOptions)
+    fs.unlinkSync(picture)
+  } else {
+    sendMessageWithBot(TELEGRAM_BOT, CHAT_ID, messageByCreated)
+    sendMessageWithBot(TELEGRAM_BOT, CHAT_ID, messageByReview)
   }
+
+  sendBodyIsValid(payload)
 }
 
 export default sendTelegram
